Move database.latest_metric() to internal.database.
